hft/marketdata/requirements.txt
alpaca-py==0.8.2
aiohttp==3.8.5
aiokafka==0.8.1
prometheus-client==0.17.1
python-json-logger==2.0.7
pydantic==2.0.3
fastapi==0.101.1
uvicorn==0.23.2
